requirements.txt
loguru
sentence-transformers
transformers>=4.6.0
pyzmq>=17.1.0
GPUtil>=1.3.0
numpy